+++ spec/Attributes/ExampleSpec.php
@@ -7,8 +7,7 @@ namespace spec\hanneskod\readmetester\Attributes;
 use hanneskod\readmetester\Attributes\Example;
 use hanneskod\readmetester\Compiler\TransformationInterface;
 use hanneskod\readmetester\Example\ExampleInterface;
-use hanneskod\readmetester\Name\NamespacedName;
-use hanneskod\readmetester\Name\NameInterface;
+use hanneskod\readmetester\Utils\Name;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
@@ -24,13 +23,13 @@ class ExampleSpec extends ObjectBehavior
         $this->shouldHaveType(TransformationInterface::CLASS);
     }
 
-    function it_transforms_name(ExampleInterface $example, NameInterface $name)
+    function it_transforms_name(ExampleInterface $example, Name $name)
     {
         $this->beConstructedWith('foo');
         $example->getName()->willReturn($name);
         $name->getNamespaceName()->willReturn('bar');
 
-        $example->withName(NamespacedName::fromString('foo', 'bar'))->willReturn($example)->shouldBeCalled();
+        $example->withName(Name::fromString('foo', 'bar'))->willReturn($example)->shouldBeCalled();
 
         $this->transform($example)->shouldReturn($example);
     }