+++ spec/Attributes/NamespaceNameSpec.php
@@ -7,8 +7,7 @@ namespace spec\hanneskod\readmetester\Attributes;
 use hanneskod\readmetester\Attributes\NamespaceName;
 use hanneskod\readmetester\Compiler\TransformationInterface;
 use hanneskod\readmetester\Example\ExampleInterface;
-use hanneskod\readmetester\Name\NamespacedName;
-use hanneskod\readmetester\Name\NameInterface;
+use hanneskod\readmetester\Utils\Name;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
@@ -26,13 +25,13 @@ class NamespaceNameSpec extends ObjectBehavior
         $this->shouldHaveType(TransformationInterface::CLASS);
     }
 
-    function it_transforms_namespace_name(ExampleInterface $example, NameInterface $name)
+    function it_transforms_namespace_name(ExampleInterface $example, Name $name)
     {
         $this->beConstructedWith('foo');
         $example->getName()->willReturn($name);
         $name->getShortName()->willReturn('bar');
 
-        $example->withName(new NamespacedName('foo', 'bar'))->willReturn($example)->shouldBeCalled();
+        $example->withName(new Name('foo', 'bar'))->willReturn($example)->shouldBeCalled();
 
         $this->transform($example)->shouldReturn($example);
     }