+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project now adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-- Nothing..
+- Removed built in phpunit integration
+- Removed the `named-only` command line option
 
 ## [1.0.0] - 2020-07-06